Make stack traces start at the task entrypoint
gdb and lldb climb up the stack to identify frames. When a function gets
called, `%ebp` gets pushed to the stack which effectively tells where
to return to. If `%ebp` is forcibly set to `0`, gdb and lldb stop their
search and the last called function is root frame.

Setting `%ebp` to 0 here ensures that the entrypoint is the root frame
for stack traces for a task instead of attaching to the stack of the
task that spawned the current task.